requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
apscheduler==3.10.4
pytz==2024.1
aiohttp==3.9.3 